requirements.txt
requests==2.25.1
sqlalchemy==1.4.5
mysql-connector==2.2.9
paramiko==2.7.2
pysftp==0.2.9